Add cloud storage for file uploads
Adds file to cloudinary storage whenever a file is uploaded to the app.
The module cloudUpload.js is added for this and the fileManager.js
module is added to delete the intermediate multer file storage during
this process.

Also updates the file download route handler response in accordance with
this new change.